+++ GraphTest/GraphIsomorphismTaggerTrianglesTest.cpp
@@ -81,10 +81,10 @@ TEST(GraphIsomorphismTaggerTrianglesTest, Diamond)
     const TaggerTriangles tagger(*graph);
     ASSERT_EQ(tagger.getGraphTag(), (Tag{1, 2, 2, 2}));
 
-    ASSERT_EQ(Single(tagger.getVertexTag(0)), 1);
-    ASSERT_EQ(Single(tagger.getVertexTag(1)), 2);
-    ASSERT_EQ(Single(tagger.getVertexTag(2)), 2);
-    ASSERT_EQ(Single(tagger.getVertexTag(3)), 1);
+    ASSERT_EQ(tagger.getVertexTag(0), 1);
+    ASSERT_EQ(tagger.getVertexTag(1), 2);
+    ASSERT_EQ(tagger.getVertexTag(2), 2);
+    ASSERT_EQ(tagger.getVertexTag(3), 1);
 }
 
 TEST(GraphIsomorphismTaggerTrianglesTest, Complete4)
